+++ src/compas_ghpython/scene/surfaceobject.py
@@ -36,9 +36,7 @@ class SurfaceObject(GHSceneObject, GeometryObject):
 
         """
         geometry = conversions.surface_to_rhino(self.geometry)
-
-        if self.transformation:
-            geometry.Transform(conversions.transformation_to_rhino(self.transformation))
+        geometry.Transform(conversions.transformation_to_rhino(self.worldtransformation))
 
         self._guids = [geometry]
         return self.guids